Add GitHub actions for compiling CodeHawk.
These are triggered on pushing new versions, and on new pull requests
being submitted. It's based on the experimental "setup-ocaml" GitHub action:
https://github.com/avsm/setup-ocaml

"setup-ocaml" really sets up opam, which is used to fetch recent
versions of ocaml, ocamlbuild, and zarith. The default appears to be
4.07 at this time, but it is possible to configure this action to test
multiple versions of ocaml. For example:
https://github.com/avsm/setup-ocaml#example-workflow

This workflow comes from an ocaml maintainer:
https://discuss.ocaml.org/t/github-actions-for-ocaml-opam-now-available/4745