Add a Shakefile.hs to compile with Shake
This build system starts from a list of ml files that define main
functions, and backtracks through module dependencies to produce the
object files necessary to build that executable.

This relies on the following assumptions:
 - ml and mli files always are present in pairs
 - No two modules have the same name

Library directories are effectively ignored, and only exist for
convenience of organization. While circular dependencies between files
are still not valid for ocamlopt, there is no requirement that build
directories have this property.

Also worth noting is that ml / mli file pairs not referenced by any
known executable are ignored and not compiled.