Make all ml and mli files come in pairs.
The *Types.ml files are symlinked to the *Types.mli files to track any
changes.

Executables are given empty .mli files.

This helps with the shake build system as it can assume that they always
come in pairs and avoid some file existence checking logic.

Unfortunately this also required adding some empty .mli files to extlib.
This shouldn't affect any of the code in CodeHawk that uses extlib,
meaning that it won't impede replacing the local copy of extlib with one
retrieved from opam.