Add opam integration
This allows compiling with --opam=SWITCH or --ocaml=VERSION. If only
"--ocaml" is specified, it will use a switch called "codehawk-VERSION".

If the switch does not exist, it will be created with the --ocaml
version if set, otherwise using a default version (currently 4.07.1).

The build system will install the libraries it needs into the current
opam switch. Changing the switch it is compiled with will invalidate the
previous build artifacts.