CHC: Close POs that refer to argv[0]
We already had some logic to deal with this but not for all pointer cases.
I copy/pasted that logic over to the cases where the current version of CHC
complains for a simple hello world program that passes argv[0] to printf.

There's still an open issue where there's a precondition for 'initialized(*argv)'
that I have not addressed