CHC: Address precondition on  'initialized(*argv)' for a simple hello world program
I couldn't figure out how to do it within the cCHCheckValid.ml match
logic so I added the logic to the actual initialized checker. It now
has a specific path that figures out if what we're looking at is
argv[0] and marks it safe, closes the PO, and does not mark it for
delegation.

Probably not the cleanest solution, but it will do for now.